Refactored algorithms into a more extensible file structure
