--- src-tauri/src/analytics/algo_result_enums/mod.rs
@@ -1,7 +1,0 @@
-pub(crate) mod ap;
-pub(crate) mod diff_cen;
-pub(crate) mod eigenvals;
-pub(crate) mod mincut;
-pub(crate) mod most_sim_timeline;
-pub(crate) mod pred_state;
-pub(crate) mod sw_cut;
