+++ src-tauri/src/analytics/algorithms/diffusion_centrality/results.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
@@ -22,6 +24,23 @@ impl Clone for DiffCenResult {
             DiffCenResult::Success(diff_cen_res) => DiffCenResult::Success(diff_cen_res.clone()),
             DiffCenResult::Error(err) => DiffCenResult::Error(err.clone()),
             DiffCenResult::Empty(empty) => DiffCenResult::Empty(*empty),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum EigenvalsResult {
+    Success(Vec<f64>),
+    Error(String),
+    Empty(bool),
+}
+
+impl Clone for EigenvalsResult {
+    fn clone(&self) -> Self {
+        match self {
+            EigenvalsResult::Success(eigenvals) => EigenvalsResult::Success(eigenvals.clone()),
+            EigenvalsResult::Error(err) => EigenvalsResult::Error(err.clone()),
+            EigenvalsResult::Empty(empty) => EigenvalsResult::Empty(*empty),
         }
     }
 }
