+++ src-tauri/src/analytics/algorithms/global_mincut/results.rs
