+++ src-tauri/src/analytics/algorithms/mod.rs
@@ -1,4 +1,6 @@
 pub mod articulation_point;
 pub mod diffusion_centrality;
 pub mod global_mincut;
+pub mod most_similar_timeline;
+pub mod predicted_state;
 pub mod stoer_wagner;
