+++ src-tauri/src/analytics/algorithms/predicted_state/mod.rs
@@ -1,0 +1,1 @@
+pub mod results;
