+++ src-tauri/src/analytics/algorithms/predicted_state/results.rs
