+++ src-tauri/src/analytics/algorithms/stoer_wagner/results.rs
